Move pipeline pipeline contracts to domain
