+++ src/bioetl/application/pipelines/error_policy_manager.py
@@ -2,11 +2,11 @@
 
 from collections.abc import Callable
 
-from bioetl.application.pipelines.hooks import ErrorPolicyABC
 from bioetl.application.pipelines.hooks_manager import HooksManager
 from bioetl.domain.enums import ErrorAction
 from bioetl.domain.errors import PipelineStageError
 from bioetl.domain.models import RunContext
+from bioetl.domain.pipelines.contracts import ErrorPolicyABC
 from bioetl.domain.providers import ProviderId
 from bioetl.infrastructure.logging.contracts import LoggerAdapterABC
