+++ src/bioetl/application/pipelines/hooks_impl.py
@@ -3,10 +3,10 @@ from __future__ import annotations
 
 from typing import Any
 
-from bioetl.application.pipelines.hooks import ErrorPolicyABC, PipelineHookABC
 from bioetl.domain.enums import ErrorAction
 from bioetl.domain.errors import PipelineStageError
 from bioetl.domain.models import StageResult
+from bioetl.domain.pipelines.contracts import ErrorPolicyABC, PipelineHookABC
 from bioetl.infrastructure.logging.contracts import LoggerAdapterABC
