+++ src/bioetl/domain/pipelines/__init__.py
@@ -1,0 +1,13 @@
+"""Pipeline domain contracts and defaults."""
+
+__all__ = [
+    "StageABC",
+    "PipelineHookABC",
+    "ErrorPolicyABC",
+]
+
+from bioetl.domain.pipelines.contracts import (  # noqa: E402,F401
+    ErrorPolicyABC,
+    PipelineHookABC,
+    StageABC,
+)
