+++ src/bioetl/domain/pipelines/contracts.py
@@ -1,6 +1,5 @@
-"""
-Pipeline lifecycle hooks.
-"""
+"""Domain-level pipeline contracts."""
+
 from abc import ABC, abstractmethod
 from typing import Any
 
@@ -8,11 +7,23 @@ from bioetl.domain.errors import PipelineStageError
 from bioetl.domain.enums import ErrorAction
 from bioetl.domain.models import StageResult
 
+__all__ = ["StageABC", "PipelineHookABC", "ErrorPolicyABC"]
+
+
+class StageABC(ABC):
+    """Абстракция стадии пайплайна."""
+
+    @abstractmethod
+    def run(self, context: Any) -> StageResult:
+        """Запускает стадию."""
+
+    @abstractmethod
+    def close(self) -> None:
+        """Освобождает ресурсы стадии."""
+
 
 class PipelineHookABC(ABC):
-    """
-    Хуки жизненного цикла пайплайна.
-    """
+    """Хуки жизненного цикла пайплайна."""
 
     @abstractmethod
     def on_stage_start(self, stage: str, context: Any) -> None:
@@ -28,9 +39,7 @@ class PipelineHookABC(ABC):
 
 
 class ErrorPolicyABC(ABC):
-    """
-    Политика обработки ошибок.
-    """
+    """Политика обработки ошибок."""
 
     @abstractmethod
     def handle(self, error: PipelineStageError, context: Any) -> ErrorAction:
