+++ tests/bioetl/application/pipelines/test_pipeline_base.py
@@ -9,13 +9,13 @@ import pandas as pd
 import pytest
 
 from bioetl.application.pipelines.base import PipelineBase
-from bioetl.application.pipelines.hooks import PipelineHookABC
 from bioetl.application.pipelines.hooks_impl import (
     ContinueOnErrorPolicyImpl,
     FailFastErrorPolicyImpl,
 )
 from bioetl.domain.errors import PipelineStageError
 from bioetl.domain.models import RunContext
+from bioetl.domain.pipelines.contracts import PipelineHookABC
 from bioetl.domain.transform.hash_service import HashService
 from bioetl.domain.transform.transformers import HashColumnsTransformer
